Add diff edit metrics
