--- evals/apps/web/src/lib/format-currency.ts
@@ -1,6 +1,0 @@
-const formatter = new Intl.NumberFormat("en-US", {
-	style: "currency",
-	currency: "USD",
-})
-
-export const formatCurrency = (amount: number) => formatter.format(amount)
