--- evals/apps/web/src/lib/index.ts
@@ -1,3 +1,0 @@
-export { formatCurrency } from "./format-currency"
-export { formatDuration } from "./format-duration"
-export { formatTokens } from "./format-tokens"
